Move utils into util package
