+++ src/main/java/com/github/dockerjava/core/command/BuildImageCmdImpl.java
@@ -10,7 +10,7 @@ import java.net.URI;
 import com.github.dockerjava.api.command.BuildImageCmd;
 import com.github.dockerjava.api.model.AuthConfigurations;
 import com.github.dockerjava.api.model.BuildResponseItem;
-import com.github.dockerjava.core.FilePathUtil;
+import com.github.dockerjava.core.util.FilePathUtil;
 import com.github.dockerjava.core.dockerfile.Dockerfile;
 
 /**
